feat(standings): add compact session footer and move weather/pitstop stats
    - Create SessionFooter component styled with a darker background and narrow padding
    - Move player pitstops and weather AIR/TRK pills to the standings footer
    - Clean up unused weather/pitstop imports and rendering in SessionHeader

Closed #189
