+++ src/widgets/StandingsWidget/SessionFooter/SessionFooter.module.scss
@@ -1,0 +1,56 @@
+.sessionFooter {
+  background: rgba(18, 18, 20, $opacity-heavy);
+  padding: ws(5) ws(10); // Narrower vertical padding (exactly like StandingsHeader)
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  border-top: 1px solid $widget-row-border;
+  font-family: $font-mono;
+  text-transform: uppercase;
+  font-size: ws(12);
+  font-weight: 400;
+  color: $widget-text-secondary;
+}
+
+.footerLeft {
+  display: flex;
+  gap: ws(8);
+  align-items: center;
+}
+
+.footerRight {
+  display: flex;
+  gap: ws(8);
+  align-items: center;
+}
+
+// Stats Pill inside footer
+.statPill {
+  display: inline-flex;
+  align-items: center;
+  gap: ws(5);
+  background: rgba(
+    0,
+    0,
+    0,
+    0.65
+  ); // Slightly darker background for footer pills
+  border: 1px solid rgba(255, 255, 255, 0.08);
+  box-shadow: inset 0 1px 3px rgba(0, 0, 0, 0.5);
+  border-radius: ws(4);
+  padding: ws(2) ws(7);
+  font-family: $font-mono;
+  font-size: ws(11);
+  letter-spacing: 0.05em;
+  line-height: 1;
+  flex-shrink: 0;
+}
+
+.statLabel {
+  color: $widget-text-secondary;
+}
+
+.statValue {
+  color: #f1f1f3;
+  font-weight: 700;
+}
