+++ src/widgets/StandingsWidget/SessionHeader/SessionHeader.tsx
@@ -1,44 +1,25 @@
 import { observer } from 'mobx-react-lite';
-import {
-  Trophy,
-  Users,
-  TriangleAlert,
-  Wrench,
-  Thermometer,
-} from 'lucide-react';
+import { Trophy, Users, TriangleAlert } from 'lucide-react';
 
 import {
   formatIRating,
   NEAR_DQ_INCIDENT_THRESHOLD,
-  getAirTempColor,
-  getTrackTempColor,
 } from '@utils/widget/widget-utils';
-import {
-  formatTemp,
-  tempUnit,
-  resolveSessionLaps,
-} from '@utils/formatters/telemetry-format';
-import {
-  computeClassSof,
-  parseWeekendTemp,
-} from '@utils/widget/standings-utils';
+import { resolveSessionLaps } from '@utils/formatters/telemetry-format';
+import { computeClassSof } from '@utils/widget/standings-utils';
 
 import type { StandingsWidgetSettings } from '@/types/widget-settings';
 import styles from './SessionHeader.module.scss';
 import {
   useBackendComputedStore,
   useTelemetryStore,
-  useUnitsStore,
   useWidgetSettingsStore,
 } from '@store/root-store-context';
 
 export const SessionHeader = observer(() => {
-  const { standings, pitStops } = useBackendComputedStore();
+  const { standings } = useBackendComputedStore();
   const telemetry = useTelemetryStore();
-
-  const { sessionInfo, weekendInfo, environment } = telemetry;
-
-  const { unitSystem } = useUnitsStore();
+  const { sessionInfo, weekendInfo } = telemetry;
   const widgetSettings = useWidgetSettingsStore();
 
   const settings =
@@ -55,7 +36,6 @@ export const SessionHeader = observer(() => {
   const playerIncidents =
     driverEntries.find((entry) => entry.isPlayer)?.incidents ?? 0;
 
-  const playerPitStops = pitStops?.playerStops ?? 0;
   const sessions = sessionInfoData?.Sessions;
   const currentSession = sessions?.[sessionInfoData?.CurrentSessionNum ?? 0];
   const trackName = weekendInfo?.TrackDisplayName ?? '';
@@ -73,24 +53,6 @@ export const SessionHeader = observer(() => {
         telemetry.leaderBestLapTime
       )
     : null;
-
-  const airCelsius =
-    environment?.air_temp ?? parseWeekendTemp(weekendInfo?.TrackAirTemp);
-
-  const trkCelsius =
-    environment?.track_temp ?? parseWeekendTemp(weekendInfo?.TrackSurfaceTemp);
-
-  const tUnit = tempUnit(unitSystem);
-
-  const airStr =
-    airCelsius !== null
-      ? `${formatTemp(airCelsius, unitSystem)}${tUnit}`
-      : null;
-
-  const trkStr =
-    trkCelsius !== null
-      ? `${formatTemp(trkCelsius, unitSystem)}${tUnit}`
-      : null;
 
   return (
     <div className={styles.sessionHeader}>
@@ -156,30 +118,6 @@ export const SessionHeader = observer(() => {
             >
               {playerIncidents}x
             </span>
-          </span>
-        )}
-
-        {settings.showPitStops && (
-          <span className={styles.statPill}>
-            <Wrench size={10} color="#9ca3af" />
-            <span className={styles.statLabel}>PIT</span>
-            <span className={styles.statValue}>{playerPitStops}</span>
-          </span>
-        )}
-
-        {settings.showWeather && airCelsius !== null && airStr && (
-          <span className={styles.statPill}>
-            <Thermometer size={10} color={getAirTempColor(airCelsius)} />
-            <span className={styles.statLabel}>AIR</span>
-            <span className={styles.statValue}>{airStr}</span>
-          </span>
-        )}
-
-        {settings.showWeather && trkCelsius !== null && trkStr && (
-          <span className={styles.statPill}>
-            <Thermometer size={10} color={getTrackTempColor(trkCelsius)} />
-            <span className={styles.statLabel}>TRK</span>
-            <span className={styles.statValue}>{trkStr}</span>
           </span>
         )}
       </div>
